# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Google Ads Service",
    "description": "Wraps the Google Ads API (v23) for MCC agency management. Handles OAuth, account linking, campaign CRUD, and performance reporting.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "serviceKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Service-to-service API key"
      }
    },
    "schemas": {
...
      "AuthUrlResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "url"
        ],
        "additionalProperties": false
      },
      "AuthCallbackResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "accountId": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "accountId",
          "message"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      }
    },
    "parameters": {
      "OrgId": {
        "name": "x-org-id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Internal org UUID from client-service"
      },
      "UserId": {
        "name": "x-user-id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Internal user UUID from client-service"
      },
      "RunId": {
        "name": "x-run-id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Caller's run ID from runs-service (used as parentRunId when creating a child run)"
      },
      "FeatureSlug": {
        "name": "x-feature-slug",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Feature slug for tracking which feature triggered the request"
      }
    }
  },
  "paths": {
...
    "/auth/url": {
      "get": {
        "summary": "Generate Google OAuth2 URL for Google Ads scope",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "name": "redirectUri",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth URL generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/callback": {
      "get": {
        "summary": "OAuth2 callback — exchanges code for refresh token and links accounts",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account linked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthCallbackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}